backend/config/urls.py
"""
URL Configuration for ft_transcendence
"""
from django.contrib import admin
from django.urls import path, include
from django.conf import settings
from django.conf.urls.static import static
from rest_framework_simplejwt.views import TokenObtainPairView, TokenRefreshView

urlpatterns = [
    # Admin
    path('admin/', admin.site.urls),

    # JWT Authentication
    path('api/token/', TokenObtainPairView.as_view(), name='token_obtain_pair'),
    path('api/token/refresh/', TokenRefreshView.as_view(), name='token_refresh'),

    # API endpoints
    path('api/users/', include('backend.apps.users.urls')),
    path('api/pong/', include('backend.apps.pong.urls')),
    path('api/chat/', include('backend.apps.chat.urls')),
    path('api/rps/', include('backend.apps.rps.urls')),
    path('api/blockchain/', include('backend.apps.blockchain.urls')),
]

# Serve media files in development
if settings.DEBUG:
    urlpatterns += static(settings.MEDIA_URL, document_root=settings.MEDIA_ROOT)
    urlpatterns += static(settings.STATIC_URL, document_root=settings.STATIC_ROOT)
