requirements.txt
# Django Core
Django==4.2.7
django-environ==0.11.2

# Database
psycopg2-binary==2.9.9

# Django REST Framework
djangorestframework==3.14.0
django-cors-headers==4.3.0

# Authentication & OAuth
djangorestframework-simplejwt==5.3.0
PyJWT==2.8.0
requests==2.31.0
oauthlib==3.2.2
requests-oauthlib==1.3.1

# WebSockets (Django Channels)
channels==4.0.0
channels-redis==4.1.0
redis==5.0.1
daphne==4.0.0

# Image handling
Pillow==10.1.0

# Blockchain (Ethereum)
web3==6.11.3
py-solc-x==2.0.2

# Security
argon2-cffi==23.1.0
python-decouple==3.8

# Utils
python-dateutil==2.8.2

# Development
ipython==8.18.1
